Kontrola - bzučák, cmd ble tunelu

+++ Kontrola_Syst/Core/Inc/flash_logger.h
@@ -2,6 +2,8 @@
 #define FLASH_LOGGER_H
 
 #include "stm32wbxx_hal.h"
+#include "app_conf.h"
+#include <stdbool.h>
 
 // Stejné adresy jako u Závodníka (zbylo nám tu 159 volných stránek)
 #define LOGGER_START_ADDR       0x08021000
@@ -22,7 +24,7 @@ typedef union {
 void Logger_Init(void);
 void Config_Init(void);
 void Logger_SavePunch_Kontrola(uint8_t* id_3bytes, uint8_t sub_sec, uint32_t unix_time);
-void Logger_GetDownloadData(uint8_t cmd, uint8_t param, uint8_t **start_ptr, uint32_t *len);
+void Logger_GetDownloadData(uint8_t cmd, uint32_t param, uint8_t **start_ptr, uint32_t *len);
 void Logger_FormatAll(void);
 void Config_Commit(BeaconConfig_t *new_cfg); // Změněno na BeaconConfig_t!
 void Config_EraseAndReboot(void);
